Optimize AST builder
