--- src/Language/ASTBuilder/VariableASTBuilder.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace Digia\GraphQL\Language\ASTBuilder;
-
-use Digia\GraphQL\Language\LexerInterface;
-use Digia\GraphQL\Language\Node\NodeKindEnum;
-use Digia\GraphQL\Language\TokenKindEnum;
-
-class VariableASTBuilder extends AbstractASTBuilder
-{
-    /**
-     * @inheritdoc
-     */
-    public function supportsBuilder(string $kind): bool
-    {
-        return $kind === ASTKindEnum::VARIABLE;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function build(LexerInterface $lexer, array $params): ?array
-    {
-        $start = $lexer->getToken();
-
-        $this->expect($lexer, TokenKindEnum::DOLLAR);
-
-        return [
-            'kind' => NodeKindEnum::VARIABLE,
-            'name' => $this->buildAST(ASTKindEnum::NAME, $lexer),
-            'loc'  => $this->buildLocation($lexer, $start),
-        ];
-    }
-}
